requirements.txt
GitPython==3.1.31
inquirer==3.1.3
